requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-ollama
langgraph
psycopg2-binary
sqlalchemy
python-dotenv
typing-extensions